get the damn modules working
